pyproject.toml
[tool.poetry]
name = "aristoph"
version = "1.0.0"
description = "School & course management platform"
authors = ["Giacomo Sarrocco"]

[tool.poetry.dependencies]
python = "^3.6"
Django = "^3.0.6"
algoliasearch-django = "^1.7.1"
django-crispy-forms = "^1.9.1"
django-import-export = "^2.1.0"
django-sendgrid-v5 = "^0.8.1"
gunicorn = "^20.0.4"
XlsxWriter = "^1.2.9"
psycopg2-binary = "^2.8.5"
xhtml2pdf = "^0.2.4"
django-markdownify = "^0.8.0"
django-extensions = "^3.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
